+++ src/main/java/data/prince.txt
@@ -1,2 +1,4 @@
 T | 0 | print task
-D | 1 | shower | Dec 21 2023, 6:00 pm
+T | 0 | park car
+T | 0 | play chess
+D | 0 | submit | Dec 31 2023, 12:00 pm
